refactors to follow a similar project structure, test and build scripts than our NPM packages
