--- index.ts
@@ -1,12 +1,0 @@
-import { EdgeConfigApiStorageWrapper } from './EdgeConfigApiStorageWrapper';
-import { EdgeConfigSdkStorageWrapper } from './EdgeConfigSdkStorageWrapper';
-import { EdgeConfigSynchronize } from './EdgeConfigSynchronize';
-export { EdgeConfigApiStorageWrapper } from './EdgeConfigApiStorageWrapper';
-export { EdgeConfigSdkStorageWrapper } from './EdgeConfigSdkStorageWrapper';
-export { EdgeConfigSynchronize } from './EdgeConfigSynchronize';
-
-module.exports = {
-  EdgeConfigSdkStorageWrapper,
-  EdgeConfigApiStorageWrapper,
-  EdgeConfigSynchronize,
-}
