+++ jest.config.js
@@ -1,8 +1,4 @@
 module.exports = {
-  roots: ['./'],
-  testMatch: ['**/__tests__/**/*.spec.ts'],
-  testPathIgnorePatterns: [
-    '<rootDir>/node_modules/',
-    '<rootDir>/dist/',
-  ],
+  preset: 'ts-jest',
+  testMatch: ['<rootDir>/src/**/__tests__/**/*.spec.ts'],
 };
