+++ package.json
@@ -1,37 +1,40 @@
 {
   "name": "@splitsoft/vercel-integration-utils",
   "version": "0.1.0",
-  "description": "",
-  "main": "dist/index.js",
+  "description": "Split SDK wrapper to integrate with Vercel Edge Config",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "types": "dist/types",
+  "files": [
+    "README.md",
+    "CONTRIBUTORS-GUIDE.md",
+    "LICENSE",
+    "CHANGES.txt",
+    "dist"
+  ],
   "scripts": {
-    "build": "tsc",
+    "build": "tsc --declaration --declarationDir dist/types && tsc -m es6 --outDir dist/esm --declaration false",
     "test": "jest"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/splitio/vercel-integration-utils.git"
   },
-  "author": "",
-  "license": "ISC",
+  "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/splitio/vercel-integration-utils/issues"
   },
   "homepage": "https://github.com/splitio/vercel-integration-utils#readme",
   "dependencies": {
-    "@splitsoftware/splitio-browserjs": "^0.9.2",
-    "@splitsoftware/splitio-sync-tools": "^0.2.0",
     "@vercel/edge-config": "0.1.4",
     "node-fetch": "^2.6.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.13",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/preset-typescript": "^7.18.6",
     "@types/jest": "^28.1.8",
     "@types/node-fetch": "^2.6.3",
     "jest": "^29.5.0",
-    "jest-fetch-mock": "^3.0.3",
+    "ts-jest": "^29.1.0",
     "typescript": "^5.0.2"
   },
-  "types": "./dist/index.d.ts"
+  "sideEffects": false
 }
