+++ src/index.ts
@@ -1,0 +1,2 @@
+export { EdgeConfigApiStorageWrapper } from './EdgeConfigApiStorageWrapper';
+export { EdgeConfigSdkStorageWrapper } from './EdgeConfigSdkStorageWrapper';
